Former/requirements.txt
asttokens==2.4.1
astunparse==1.6.3
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
fonttools==4.49.0
importlib_resources==6.1.3
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.5
mindspore @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/2.2.11/MindSpore/unified/x86_64/mindspore-2.2.11-cp38-cp38-linux_x86_64.whl#sha256=b6696b8273cb950b2d6ed36bda40f78f92abdb473f795fa6f49a4ab44edfb36f
mindspore-hub @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/1.9.0/Hub/any/mindspore_hub-1.9.0-py3-none-any.whl#sha256=56b982ea2f7afa1b3334d6edd4d341879e74ba8b7263ee82657f7a81b31f3b8c
mistune==0.8.4
numpy==1.24.4
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.3.0
zipp==3.17.0
